+++ crates/wasi-stub/src/lib.rs
@@ -9,10 +9,12 @@ use wast::{
     Wat,
 };
 
+#[derive(Debug)]
 pub enum FunctionsToStub {
     All,
     Some(HashSet<String>),
 }
+#[derive(Debug)]
 pub struct ShouldStub {
     pub modules: HashMap<String, FunctionsToStub>,
 }