+++ CHANGELOG.md
@@ -53,6 +53,9 @@ All dates in this file are in UTC.
 - Support stubbing functions that return `i64`, `f32`, or `f64` — only `i32` was supported previously (#32).
 - Support [cargo-binstall](https://crates.io/crates/cargo-binstall) (#59).
 
+#### Changed
+- **(Breaking)** The WASI module will no longer be stubbed by default if `--stub-module` or `--stub-function` is given, making it possible to only stub some of the WASI functions; you can pass `--stub-module wasi_snapshot_preview1,other_module` to restore the old behavior (#17).
+
 ### [0.2.1] - 2026-03-02
 
 This version is mainly for testing [trusted publishing](https://crates.io/docs/trusted-publishing) after the repository has been [transferred from astrale-sharp to typst-community](https://github.com/orgs/typst-community/discussions/35).